+++ src/main/java/in/koreatech/koin/admin/banner/controller/AdminBannerController.java
@@ -1,13 +1,16 @@
 package in.koreatech.koin.admin.banner.controller;
 
 import static in.koreatech.koin.domain.user.model.UserType.ADMIN;
+import static io.swagger.v3.oas.annotations.enums.ParameterIn.PATH;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,9 +18,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 import in.koreatech.koin._common.auth.Auth;
 import in.koreatech.koin.admin.banner.dto.request.AdminBannerCreateRequest;
+import in.koreatech.koin.admin.banner.dto.request.ChangeBannerActiveRequest;
+import in.koreatech.koin.admin.banner.dto.request.ChangeBannerPriorityRequest;
+import in.koreatech.koin.admin.banner.dto.request.ModifyBannerRequest;
 import in.koreatech.koin.admin.banner.dto.response.AdminBannerResponse;
 import in.koreatech.koin.admin.banner.dto.response.AdminBannersResponse;
 import in.koreatech.koin.admin.banner.service.AdminBannerService;
+import io.swagger.v3.oas.annotations.Parameter;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
@@ -65,5 +72,35 @@ public class AdminBannerController implements AdminBannerApi {
     ) {
         adminBannerService.deleteBanner(bannerId);
         return ResponseEntity.noContent().build();
+    }
+
+    @PatchMapping("/{id}/priority")
+    public ResponseEntity<Void> changePriority(
+        @Parameter(in = PATH) @PathVariable Integer id,
+        @RequestBody @Valid ChangeBannerPriorityRequest request,
+        @Auth(permit = {ADMIN}) Integer adminId
+    ) {
+        adminBannerService.changePriority(id, request);
+        return ResponseEntity.ok().build();
+    }
+
+    @PatchMapping("/{id}/active")
+    public ResponseEntity<Void> changeActive(
+        @Parameter(in = PATH) @PathVariable Integer id,
+        @RequestBody @Valid ChangeBannerActiveRequest request,
+        @Auth(permit = {ADMIN}) Integer adminId
+    ) {
+        adminBannerService.changeActive(id, request);
+        return ResponseEntity.ok().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Void> modifyBanner(
+        @Parameter(in = PATH) @PathVariable Integer id,
+        @RequestBody @Valid ModifyBannerRequest request,
+        @Auth(permit = {ADMIN}) Integer adminId
+    ) {
+        // bannerService.modifyBanner(id, request);
+        return ResponseEntity.ok().build();
     }
 }
