+++ src/main/java/in/koreatech/koin/admin/banner/dto/request/ChangeBannerActiveRequest.java
@@ -1,6 +1,6 @@
-package in.koreatech.koin.domain.banner.dto;
+package in.koreatech.koin.admin.banner.dto.request;
 
-import static com.fasterxml.jackson.databind.PropertyNamingStrategies.*;
+import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
