+++ src/main/java/in/koreatech/koin/admin/banner/dto/request/ChangeBannerPriorityRequest.java
@@ -1,11 +1,11 @@
-package in.koreatech.koin.domain.banner.dto;
+package in.koreatech.koin.admin.banner.dto.request;
 
-import static com.fasterxml.jackson.databind.PropertyNamingStrategies.*;
-import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.*;
+import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
+import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
-import in.koreatech.koin.domain.banner.enums.PriorityChangeType;
+import in.koreatech.koin.admin.banner.enums.PriorityChangeType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
