+++ src/main/java/in/koreatech/koin/admin/banner/dto/request/ModifyBannerRequest.java
@@ -1,6 +1,6 @@
-package in.koreatech.koin.domain.banner.dto;
+package in.koreatech.koin.admin.banner.dto.request;
 
-import static com.fasterxml.jackson.databind.PropertyNamingStrategies.*;
+import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
