--- src/main/java/in/koreatech/koin/domain/banner/repository/BannerCategoryRepository.java
@@ -1,9 +1,0 @@
-package in.koreatech.koin.domain.banner.repository;
-
-import org.springframework.data.repository.Repository;
-
-import in.koreatech.koin.domain.banner.model.BannerCategory;
-
-public interface BannerCategoryRepository extends Repository<BannerCategory, Integer> {
-
-}
